Fix #110 #110 - Related fix: set the desired state in the zone after a stop/start action, in order to sooner reflect it on screen #110 - Related fix: fix the counter timer start/stop logic (to not run in the background when not necessary) #110 - Minor related fix: round the counter to minutes better (7:25 —> 7, 7:35 —> 8) NSDictionary+Keychain: Minor cast fixes

+++ Sprinklers/_Additions/NSDictionary+Keychain.m
@@ -30,7 +30,7 @@
                                        };
         OSStatus osStatus = SecItemAdd((CFDictionaryRef)storageQuery, nil);
         if(osStatus != noErr) {
-            DLog(@"Keychain store error: %d", osStatus);
+            DLog(@"Keychain store error: %d", (int)osStatus);
         }
     }
 }
@@ -81,7 +81,7 @@
         // do delete
         osStatus = SecItemDelete((CFDictionaryRef)deleteQuery);
         if(osStatus != noErr) {
-            DLog(@"Keychain delete error: %d", osStatus);
+            DLog(@"Keychain delete error: %d", (int)osStatus);
         }
         [deleteQuery release];
     }
